+++ Frends.FTP.DownloadFiles/Frends.FTP.DownloadFiles.Tests/SourceFilesNotFoundActionTests.cs
@@ -69,12 +69,26 @@ public class SourceFilesNotFoundActionTests : DownloadFilesTestBase
             NotFoundAction = sourceNotFoundAction
         };
         var destination = new Destination
-        { Directory = targetDir, Action = DestinationAction.Overwrite };
+        {
+            Directory = targetDir,
+            Action = DestinationAction.Overwrite
+        };
         var connection = FtpHelper.GetFtpConnection();
 
         var result = FTP.DownloadFiles(
-            source, destination, connection,
-            new Options { OperationLog = true }, new Info(), new CancellationToken());
+            new Input
+            {
+                Source = source,
+                Destination = destination,
+                Info = new Info()
+            },
+            connection,
+            new Options
+            {
+                OperationLog = true,
+                ThrowErrorOnFailure = false
+            }, new CancellationToken());
+
         return result;
     }
 }